+++ startup/gui/graphEditor.py
@@ -38,6 +38,7 @@
 import functools
 import pathlib
 import re
+import weakref
 
 import imath
 
@@ -185,11 +186,11 @@ def __canDropFiles( graphGadget, event ) :
 		not Gaffer.MetadataAlgo.getChildNodesAreReadOnly( graphGadget.getRoot() )
 	)
 
-def __dragEnter( graphGadget, event ) :
+def __fileDragEnter( graphGadget, event ) :
 
 	return __canDropFiles( graphGadget, event )
 
-def __drop( graphGadget, event ) :
+def __fileDrop( graphGadget, event ) :
 
 	if not __canDropFiles( graphGadget, event ) :
 		return False
@@ -233,9 +234,90 @@ def __drop( graphGadget, event ) :
 
 	return True
 
+##########################################################################
+# Scene location drop handler
+##########################################################################
+
+def __dropLocationData( event ) :
+
+	if (
+		not isinstance( event.data, IECore.StringVectorData ) or
+		len( event.data ) != 1 or
+		not event.data[0].startswith( "/" )
+	) :
+		return None
+
+	scene = None
+	sourceEditor = event.sourceWidget.ancestor( GafferUI.Editor )
+	if isinstance( sourceEditor, GafferUI.Viewer ) :
+		if isinstance( event.sourceGadget, GafferSceneUI.SceneGadget ) :
+			scene = sourceEditor.view()["in"].getInput()
+	elif isinstance( sourceEditor, GafferSceneUI.HierarchyView ) :
+		scene = sourceEditor.scene()
+
+	if scene is None :
+		return None
+
+	return {
+		"path" : event.data[0],
+		"scene" : scene,
+		"context" : sourceEditor.getContext(),
+	}
+
+def __locationDragEnter( graphGadget, event ) :
+
+	if __dropLocationData( event ) is None :
+		return False
+
+	GafferUI.Pointer.setCurrent( "target" )
+	return True
+
+def __locationDragLeave( graphGadget, event ) :
+
+	if __dropLocationData( event ) is not None :
+		if event.destinationWidget is None :
+			# Hack to restore (what we assume to have been) the original
+			# drag pointer. We don't do this when another widget has
+			# accepted the drag, because it would clobber any pointer
+			# change they made in `dragEnter`. But of course that means
+			# that if another widget _has_ accepted the drag but hasn't
+			# changed the pointer themselves (maybe they use highlighting
+			# instead), they will be stuck with the wrong pointer.
+			## \todo This is far too fragile. We need to manage
+			# pointer restoration at a higher level, in Widget.py's
+			# _EventFilter (and ViewportGadget's handlers).
+			GafferUI.Pointer.setCurrent( "objects" )
+		return True
+
+	return False
+
+def __locationDrop( graphGadget, event, graphEditor ) :
+
+	dropLocationData = __dropLocationData( event )
+	if dropLocationData is None :
+		return False
+
+	with dropLocationData["context"] :
+		sourceScene = GafferScene.SceneAlgo.source( dropLocationData["scene"], dropLocationData["path"] )
+
+	if sourceScene is not None :
+		graphGadget.setRoot( sourceScene.node().parent() )
+		## \todo The `frame()` method should probably be on the GraphGadget itself, and the `at`
+		# functionality should be made public.
+		graphEditor()._GraphEditor__frame(
+			[ sourceScene.node() ],
+			at = imath.V2f( GafferUI.Widget.mousePosition( relativeTo = graphEditor() ) )
+		)
+
+	return True
+
 def __graphEditorCreated( graphEditor ) :
 
-	graphEditor.graphGadget().dragEnterSignal().connect( __dragEnter, scoped = False )
-	graphEditor.graphGadget().dropSignal().connect( __drop, scoped = False )
+	graphEditor.graphGadget().dragEnterSignal().connect( __fileDragEnter, scoped = False )
+	graphEditor.graphGadget().dropSignal().connect( __fileDrop, scoped = False )
+
+	graphEditor.graphGadget().dragEnterSignal().connect( __locationDragEnter, scoped = False )
+	graphEditor.graphGadget().dragLeaveSignal().connect( __locationDragLeave, scoped = False )
+	graphEditor.graphGadget().dropSignal().connect( functools.partial( __locationDrop, graphEditor = weakref.ref( graphEditor ) ), scoped = False )
 
 GafferUI.GraphEditor.instanceCreatedSignal().connect( __graphEditorCreated, scoped = False )