+++ Changes.md
@@ -13,6 +13,7 @@ Improvements
   - Improved performance ~3x.
   - Improved cancellation responsiveness.
 - 3Delight : Added NSI screen static sampling pattern option (`dl:staticsamplingpattern`).
+- GraphEditor : The source node for any location can be located by dragging the location into the GraphEditor from the Viewer or HierarchyView.
 
 Fixes
 -----